pdf2txt GUI version
This is the completed first version of PDF2TXT Graphical edition. This java program and wrapper combined allow the user to run pdf2txt completely graphically, including automating installation and generation of the necessary file structure.

The user needs to have python installed for pdf2txt to work, and this program DOES NOT install python.